refactor: Put aggregator components in own folder
