+++ client/src/lib/Sources/Aggregators/FeedBulletPoint.svelte
